Give each type a file named after it
Eleven files declared a type and then some others alongside it, so finding
Level or RestrictionBreach meant knowing which producer or interface it had
been tucked behind. Thirteen types move to files of their own, in the same
folders and namespaces - nothing here is breaking.

TimeProvider.cs held ITimeProvider and nothing else, which was the one
outright filename/type mismatch in the codebase.

FullBookDataProducer declared its enum and its event record above the class
the file is named for; the producer now leads its own file. Every file in
src now declares its namesake first.

SelfMatchPrevention leaves OrderBookAction.cs because it is not one: it
derives from nothing and is a qualifier CreateOrder carries, which the file
now says out loud rather than leaving to be inferred from the absence of a
base type.

The four genuine hierarchies stay whole - OrderBookAction, OrderBookEvent,
MatchOutcome and PriceRestrictionConfig each describe one closed set of
alternatives, and splitting those across sixteen files would cost more than
it bought.

ISessionProvider drops its using: the only thing in it that knew about
OrderBookStatus was SessionStatusChangedArgs, which now lives elsewhere.
