+++ src/Circus/DataProducers/IndicativePriceDataEvent.cs
@@ -1,0 +1,3 @@
+namespace Circus.DataProducers;
+
+public record IndicativePriceDataEvent(DateTime Time, decimal? Price, int Quantity);
