+++ src/Circus/DataProducers/Level.cs
@@ -1,0 +1,4 @@
+namespace Circus.DataProducers;
+
+// Aggregated market data, not a book structure: one publishable line per price.
+public record Level(decimal Price, int Quantity, int Count);
