+++ src/Circus/DataProducers/LevelsDataEvent.cs
@@ -1,0 +1,3 @@
+namespace Circus.DataProducers;
+
+public record LevelsDataEvent(DateTime Time, IReadOnlyList<Level> Bids, IReadOnlyList<Level> Offers);
