+++ src/Circus/DataProducers/OrderBookDeltaAction.cs
@@ -1,0 +1,9 @@
+namespace Circus.DataProducers;
+
+public enum OrderBookDeltaAction
+{
+    Added,
+    Modified,
+    Removed,
+    Filled
+}
