+++ src/Circus/DataProducers/OrderBookDeltaEvent.cs
@@ -1,0 +1,6 @@
+namespace Circus.DataProducers;
+
+// ExchangeOrderId only - never CompanyId/ClientOrderId, which identify the originating
+// client and must not be broadcast on a public depth feed.
+public record OrderBookDeltaEvent(DateTime Time, Side Side, string ExchangeOrderId, decimal Price, int Quantity,
+    OrderBookDeltaAction Action);
