+++ src/Circus/DataProducers/TradedDataEvent.cs
@@ -1,0 +1,3 @@
+namespace Circus.DataProducers;
+
+public record TradedDataEvent(DateTime Time, decimal Price, int Quantity);
