+++ src/Circus/OrderBook/Actions/OrderBookAction.cs
@@ -43,16 +43,6 @@ public abstract record OrderAction : OrderBookAction
     public required string ClientOrderId { get; init; }
 }
 
-// Id is required: an instruction with no id is meaningless (nothing to match against), so
-// rather than let that combination be constructed and silently ignored, opting into self-match
-// prevention at all means supplying an id - Instruction is the only genuinely optional part,
-// falling back to CancelResting when omitted.
-public sealed record SelfMatchPrevention
-{
-    public required string Id { get; init; }
-    public SelfMatchPreventionInstruction? Instruction { get; init; }
-}
-
 public abstract record CreateOrder : OrderAction
 {
     public required OrderValidity OrderValidity { get; init; }
