+++ src/Circus/OrderBook/Matching/Allocation.cs
@@ -1,0 +1,3 @@
+namespace Circus.OrderBook.Matching;
+
+internal readonly record struct Allocation(InternalOrder Resting, int Quantity, long PriceTicks);
