+++ src/Circus/OrderBook/Restrictions/RestrictionBreachAction.cs
@@ -1,0 +1,9 @@
+namespace Circus.OrderBook.Restrictions;
+
+internal enum RestrictionBreachAction
+{
+    Reject,
+    Block,
+    Pause,
+    Halt
+}
