+++ src/Circus/TimeProviders/ITimeProvider.cs
